+++ src/module/ChatMessageMidi.ts
@@ -1,10 +1,34 @@
 import { GameSystemConfig, debugEnabled, i18n, log, warn } from "../midi-qol.js";
 import { configSettings } from "./settings.js";
+import { getDamageType } from "./utils.js";
 
 export class ChatMessageMidi extends globalThis.dnd5e.documents.ChatMessage5e {
   constructor(...args) {
     super(...args);
     if (debugEnabled > 1) log("Chat message midi constructor", ...args)
+  }
+
+  _aggregateDamageRoll(roll, breakdown) {
+    for ( let i = roll.terms.length - 1; i >= 0; ) {
+      const term = roll.terms[i--];
+      if ( !(term instanceof NumericTerm) && !(term instanceof DiceTerm) ) continue;
+      const flavor = term.flavor?.toLowerCase();
+      const damageType = flavor === "" ? undefined : getDamageType(flavor);
+      const type = damageType ? damageType : roll.options.type;
+      const aggregate = breakdown[type] ??= { total: 0, constant: 0, dice: [] };
+      const value: number = Number(term.total ?? 0);
+      if ( term instanceof DiceTerm ) aggregate.dice.push(...term.results.map(r => ({
+        result: term.getResultLabel(r), classes: term.getResultCSS(r).filterJoin(" ")
+      })));
+      let multiplier = 1;
+      let operator = roll.terms[i];
+      while ( operator instanceof OperatorTerm ) {
+        if ( operator.operator === "-" ) multiplier *= -1;
+        operator = roll.terms[--i];
+      }
+      aggregate.total += value * multiplier;
+      if ( term instanceof NumericTerm ) aggregate.constant += value * multiplier;
+    }
   }
 
   collectRolls(rolls) {